Frame Analysis: Add analyse_options=hold
Sometimes it may be desirable to capture several subsequent frames
during frame analysis. For these cases, add analyse_options=hold, which
will continue the frame analysis on subsequent frames so long as the
analyse key is held down.

If the option is enabled the filename and logs will include the frame
number, and the draw call count will be reset on each frame.

Unlike most other analyse_options, we don't allow this to be changed
mid-frame as it could easily lead to filename conflicts.