Restructure constant buffer initializer for default values.
This is the start of a fix for large initializer, and just restructures the code to make it handle each case specially, like we do in other place, using if then else.

This adds an else clause to just output the bad values, without initing them, which should allow the shader through at least.

Adds some convenience routines to clean up some of the overly verbose code.

Should have no impact on any shader generation, but now allow through the large initializers.