Hooks: Add except_set_samplers to work around more crashes
If these calls are hooked we crash early in MGSGZ on Windows 10. There
are similar reports of crashes on Windows 8, and Windows 7 *without* the
evil update.

Unlike the SetShaderResource crash, we can't work around this by
skipping the original call if NumSamplers=0, because this crash occurs
on seemingly valid calls, such as:

HookedContext::VSSetSamplers(StartSlot:0, NumSamplers:16, ppSamplers:0x000001DBB99D9D10)

Disabling these calls was based on a haunch as they have been implicated
in the stack of crashes in Witcher 3. In those cases the stack does not
make much sense as ExecuteCommandList had been called earlier and we
should not see a call back into these unless the command list was being
emulated in software. Although this does not make sense, the crashes
very consistently give the same stack trace (with one of the SetSampler
calls), so it is not random.

In the Witcher 3 crashes, the instruction pointer is off executing
garbage in la-la land, always a bunch of 0x90909090s following a sled of
NULLs. I cannot tell if the MGSGZ crashes are the same.

This is looking more and more like edge cases that Deviare is not
handling correctly, though more investigation is required to confirm and
determine the exact nature of the crashes.