Add hack to get both MGSV games running on Win 10
On Windows 10 these games seem more picky than they are on Windows 7,
and also seem to reject our wrapped DXGIFactory. However, they don't
reject our wrapped DXGIFactory1 and that is what they use to create the
swap chain, so rather than hook the DXGIFactory let's add a hack to not
wrap the DXGIFactory0 as the path of least resistance.

If this turns out to be insufficient we can always add full hooking of
the factory later.