Hooks: Work around yet another hang in RSSetState without the evil update
This is yet another case where we hang, but this time only on Windows 7
without the evil update installed. This time, the hang occurs the first
time a non-NULL value is passed into the function.

Hooking seems pretty fragile - I can only assume that Deviare is not
fixing up branches that are jumping mid-way into the patched
instructions (which is technically quite difficult as there is no easy
way to tell where all possible affected instructions are short of
running a code analysis across the entire program, but even that is
insufficient as is evident by the fact that IDA does that and cannot
reliably detect all code sections), or is not correctly fixing
up or emulating all instructions in the trampoline.

Notably we aren't on the latest version so it is possible some of these
issues have been addressed, but I had issues trying it as well.