Hooks: Add a hook=recommended option
Requiring so many workarounds is disappointing. I don't want to kill the
options until they are adequately resolved, but I also don't want people
to have to list a million options without a good idea of what does what.

Therefore, add a recommended option that enables all hooks except those
known to cause hangs or crashes.

This option does not skip wrapping the DXGI factory as that is necessary
to get the present call in some games, so that workaround must be
manually specified.