Ini Handler: Don't abort when parsing sections with missing key/hash
A TextureOverride or ShderOverride section with a missing hash or a Key
section with a missing Key would cause the ini parser to stop processing
further sections of the same type, and would not warn that this had
happened. This could lead to confusion as sections would seem to have no
effect, so add a warning and continue processing other sections instead
of stopping.