Frame Analysis: Add newlines when logging certain NULL objects
There's a few places where the initial frame analysis logging call does
not insert a newline if an additional piece of information is about to
be logged that should be appended to the line, on the assumption that
frame analysis will add the newline character. However, if a NULL
pointer is passed to frame analysis it will return immediately and won't
have added the newline.

This patch fixes several instances of this so that a newline will always
be added even when called with NULL.