Frame Analysis: Print command list handle in FinishCommandList
In general I tried not to dereference any pointers in most of the one
shot frame analysis calls to avoid the risk of missing a check and
crashing, but this limits the usefulness of certain functions that take
a pointer to a pointer as it may not be possible to cross-reference
elsewhere.

FinishCommandList is one such case where we need the handle of the
command list so we can match it to the corresponding ExecuteCommandList
call in the immediate context.