Separate the texture description hash from the texture data hash
For games like Far Cry 4 and Metal Gear Solid V where we need to track
updates to textures in order to keep their hash synchronised with the
contents of the texture we currently use track_texture_updates=1.

While this does solve the problem, it is rather expensive due to the way
our texture hashes are designed as it has to hash every texture updated
via Map()/Unmap() or UpdateSubresource() before it knows whether it was
necessary or not, by which time it has already spent the CPU cycles.

We can improve this by redesigning the way hashes work to separate out
the description hash from the data hash. Then we only need to rehash the
data if the description hash is used in a TextureOverride section. This
commit does not introduce that optimisation, but it makes the changes to
the hash as a precursor to doing so.

Since 3DMigoto 1.2.x we have used a CRC32 hash, which was calculated by
combining the data and description together. This change switches back
to a 64bit hash, which is made up of the data hash in the high order
32bits and the description hash in the low order 32bits.

If a hash only consisted of a description hash it will not change (other
than some extra 0s being prepended), but if it included a data hash it
will inevitably change, so this bumps the version number to 1.3.x

We could potentially add a backwards compatibility mode, but DHR was not
concerned since fixes rarely get an update to the DLL, and any game I
might update will need the new hashes anyway.