Add old_orig_hash to ShaderUsage.txt
To ease finding the new hashes when porting a fix from 3DMigoto 1.2.x to
1.3.x, add the old hash in the ShaderUsage.txt so that one only has to
have the textures displayed while hunting is enabled and dump out a
ShaderUsage.txt and search for the old hashes to find the new ones.

This does not track updates to the old hashes with
track_texture_updates, but it should still save time.