Skip hashing resource data when not necessary
Take note of each description hash mentioned in the d3dx.ini file, and
use them to eliminate the need to calculate a data hash of resources
that have no matching description in the d3dx.ini.

This is mostly aimed at reducing the cost of track_texture_updates as it
will be able to skip rehashing many resources that are updated each
frame if their descriptions do not appear in any of the textures we need
to track.

It will also reduce the cost of creating textures/buffers in release
mode as it can also skip unnecessary data hashes there.

This optimisation is disabled when in hunting mode as we will always
need the data hashes in that mode to identify new hashes.