FC4: Update d3dx.ini to use new hashes
Most of these were straight forward, but while updating these I found
that the hashes for the repair tool and rope no longer matched. At
first it seemed to just be that the texture quality settings were
selecting different textures, but that cannot explain all the symptoms
I'm seeing by itself.

Either there is a bug in the hash tracking code, or this game is copying
between different subresources which the tracking code does not handle.