Don't force debug=1 and unbuffered=1 in debug builds
We could be using a debug build for all sorts of reasons, and we don't
necessarily want unbuffered debug logging enabled, and always enabling
it can make it extremely slow to launch a game and get it to whatever we
are actually interested in. It also can interfere with race conditions
we might be trying to debug.

It's trivial to turn that on in the ini file if we do want it, so don't
force it.