Resource Copying: Make resources typeless & manually populate view
There are a few situations where creating a fully typed resource is the
wrong answer. Depth/stencil resources need their type changed to suit
the slot they are bound in the pipeline, which we were previously doing
by assuming they will be bound in a shader resource slot, but that would
not work if they were then bound in a DSV slot (and while we could check
the bind flags, there is no answer if it was to be bound in both).

It also turns out that the format conversion we were doing for the
resource was preventing us from creating copies of MSAA depth/stencil
resources with an error from the debug layer that it was not supported
by the hardware for the given format. This was likely an issue due to
using only a partially typed format for these, and using either fully
typeless or fully typed should be ok.

Instead, create untyped resources and set the type in the view. This
unfortunately means that we cannot pass NULL for the description when
creating a view, so we have to populate it ourselves, which means a lot
of boiler plate code and a fair potential for bugs to have crept in.