Allow ShaderOverride sections to override the shader model
If a shader was originally shader model 4 it restricts what we can do in
the shader. e.g. we cannot use MSAA textures in shader model 4 vertex
shaders, which can be problematic for the auto crosshair.

Add a model option to ShaderOverride sections which can be used to
override the shader model to allow us to use more recent features as
needed.