Resource copying: Implement copy_desc operation
This operation is will create (or recreate) a resource to match the
description of another existing resource (except for all the usual
changes we make, like bind flags, format, etc), but without actually
copying that resource. It is intended to be used to create resources
that can be used as a simultaneous render target, ensuring that their
width, height, MSAA, etc matches, but without clearing it every draw
call.

The implementation is very simple since the code already exists - all it
does is skip the actual copy operation.