Implement arbitrary resource creation & overriding
This extends CustomResource sections to include various options to allow
custom resources to be created from scratch rather than being created
through copying an existing resource or loading from file. The same
options will also override the resource parameters if it is created
through copying another resource.

This opens up a lot of possibilities for advanced fixes that need
additional information passed between shaders. It could be used to
create a structured buffer to pass the position of enemies to match up
with health bars, to allow matrices to be copied from one shader to
another if some conditional logic is required (e.g. to match render
targets in DreadOut), or a Texture2D could be used to pass additional
per-pixel data from a shader to another.