Injected Shaders: Add a draw=from_caller mode
We might sometimes want to run an injected shader using the same draw
call the game used as an alternative to adding additional render targets
(but with the advantage that the render targets will not need to be
compatible with any others the game is using).

This adds a draw=from_caller mode to use the same draw call and
parameters as the game. Currently does not support DrawAuto or Indirect
calls.