Update StereoParams on Present, not ClearRenderTarget
Not all games call ClearRenderTarget on a regular basis (e.g. The
Witness almost never calls it on low quality settings), which allows the
stereo parameters to easily get out of date. The use of
ClearRenderTarget was left over from when we had removed DXGI support
and is no longer necessary to do there, so move the update to
RunFrameActions to ensure that it is always up to date.

This should also reduce the number of calls into nvapi for games that
clear many render targets per frame (Frozenbyte noted that calling nvapi
too often in the Trine games had a performance impact).