Fix compiler warning that crept in with draw=from_caller
ParseDrawCommand checks to ensure that the entire value from the ini
file has been consumed in common code, and bails otherwise. For the
draw=from_caller case I set the read length to the end of the string
since wcscmp has already given this guarantee, but the return type of
string::length is different to %n used in the scanf functions. Cast it
to remove the warning.