Hook SetWindowPos on demand if full_screen=2
Oomek reported that the recent update caused issues exiting full screen:

https://forums.geforce.com/default/topic/685657/3d-vision/3dmigoto-now-open-source-/post/4801159/#4801159

While unconfirmed, a likely explanation is that the trampoline for the
SetWindowPos hook is not functioning as expected, and given the
experiences with the hooked context it seems the safest course of action
is to only hook it if we require it.

Therefore, change full_screen to use two possible values - 1 to just
force the swap chain to full screen, and 2 to also hook SetWindowPos.
Delay the hooking of SetWindowPos until we actually need to start
forcing full screen when a swap chain is created. Since this is delayed
to a time when we can use regular logging, change it to do so.

Change the toggle_full_screen to negate the full screen value so that
the current setting will be preserved when it is toggled back on, and
change the code that checks the full screen setting to check for
positive values.

Reported-by: Oomek