Ensure mResources is protected by the critical section
The mResources map is not thread safe against simultaneous modifications
and reads, which can lead to a rare crash in the resource hash tracking
code, or with hunting enabled (I have seen this occur in The Witness).

Add the critical section around it where it was missing, and expand the
protected code where it was already used elsewhere in a function.
GetOrigResourceHash and GetResourceHash are called from a number of
functions that already obtain the critical section, so document them
that the caller should obtain it and do so as necessary.

For frame analysis, just grab it around the entire section of code that
dumps resources.

With this much locking I am somewhat tempted to implement a lock free
map instead, especially since we never remove entries from the map...
but that is something that is easy to mess up, so let's see how well
this performs first.