cmd_Decompiler: Hook up Flugan's assembler and add signature parser
Flugan's assembler normally cheats and reuses sections from the original
binary when replacing a shader from the game, but that restricts what
modifications we can do and is not an option when assembling a
stand-alone shader.

In order to overcome this, implement a parser in cmd_Decompiler that
parses the comment blocks the disassembler produces for these sections
and manufactures a new shader binary with these sections. That is passed
to Flugan's assembler which will then fill out the SHDR / SHEX section.
This allows us to assemble a shader without requiring the original
binary.

This should undergo some further testing before hooking it up in game.
Hull and domain shaders lose an additional comment block, but it appears
that it comes from the RDEF section which we believe is optional
(confirmed based on the fact that fxc /dumpbin /Qstrip_reflect also
strips this).

Tested with a random selection of domain, hull, geometry, compute,
vertex and pixel shaders from Witcher 3 and Stranded Deep, as well as
a number of hand crafted shaders using various semantics.

This even handles some DX12 specific syntax (though evidently Flugan's
assembler may need some work in this area).