cmd_Decompiler: Clarify SPRs that only ever show up in the output signature
I was worried there might be an edge case I'd missed, as some of the
semantics that use special purpose registers seemed to be write-only
despite being documented as read/write, and if they could be declared in
an input signature the used mask would have to be inverted.

Investigation suggests that these do not appear in the input signature
when they are used as an input, and that their SPR name in the assembly
starts with a 'v' instead of an 'o', so it seems there is nothing to
worry about.

SV_Coverage is confirmed. SV_Depth* and the DX12 semantics are not.