Fix crash regression in stero2mono copy support
The commit "Implement arbitrary resource creation & overriding"
(ebd705474999346467f644601e32caf9e4cc2412) introduced a regression in
the stereo2mono copy mode which would crash on a NULL pointer
dereference.

For now, don't allow resource sections to override the description of
the intermediate stereo2mono resource as it has some specific
requirements. I might need to revisit that decision if it turns out that
we ever need to override the format of the intermediate resource for
whatever reason.