resource copying: Add 'bb' target to access back buffer
Drawing directly to the screen from the [Present] call only works if the
back buffer is currently assigned as a render target, and it turns out
that games like MGSV don't do that, which complicates using custom
resources for debugging effects in those games - it is still possible,
but it requires some back tracking in the frame analysis log to find the
last shader used in the frame that did render to the back buffer and
injecting the custom shader there.

Add the ability to get access to the back buffer in the arbitrary
resource copying support to allow for a universal solution to this
problem.