Tweaks to present command list & overlay, add post-present command list
If we are running commands from the present call, it is conceivable that
we might need to save or restore some state across the call, or perform
actions explicitly before or after the present call, so add a
post-present command list. This also makes the present section more
consistent with other command list enabled sections as the post keyword
will behave the same.

Move the two command lists around a little so that state is correct for
each - the pre-present command list is run before processing input or
updating the frame count, and the post present call is run after all
that and after the StereoParams resource has been updated.

Additionally, if we are running a custom shader off the present call to
render something to the back buffer we would draw over the overlay,
which does not seem desirable so move where the overlay is drawn until
after the pre-present command list has been run.