Command list: Fix bug adding commands to wrong command list
If a general command was explicitly added to a pre or post command list
it would cause all subsequent general commands without an explicit list
to be added to the same list rather than the default, e.g.

[Present]
post ResourceDebug = bb
run = CustomShaderDebug

In this example ResourceDebug = bb is explicitly added to the post
command list, while run = CustomShaderDebug was supposed to be added to
the pre command list (default for run commands), but got added to the
post command list instead.

The fix is simple - set the explicit command list to NULL in each
iteration of the loop before parsing it.