Fix bug where StereoParams were not injected into hull, domain & geometry shaders
Commit "Bind StereoParams and IniParams on draw calls instead of
SetShader" (9ab123c42bcbf0d4f96be71f10ad34127178fb6e) changed when the
StereoParams and IniParams were bound to the draw call, however for
hull, domain & geometry shaders it checked if a shader was bound before
trying to bind the resources as these types of shaders rarely are.

Unfortunately this was caught out by an optimisation that leads to the
current shaders not being updated in certain circumstances, so these
shaders would sometimes not have StereoParams and IniParams bound at
all.

Always update the current shader handle regardless as that should be
cheap, and use that to test if these shaders are bound instead of the
hash.