Arbitrary resource copying: Add a set_viewport keyword
Some games like MGSV don't have a valid view-port assigned on the
present call, which we need to be able to inject shaders to run on it.
Add a set_viewport keyword to set the viewport to the dimensions of the
assigned resource, e.g.

o0 = set_viewport bb

There is no way to save & restore the viewport from the game yet