Custom Shaders: Add ability to override the rasterizer state
When we inject a custom shader we are subject to whatever state the game
has left for us, which may not be suitable for our purposes. Even a
simple full screen shader may be impacted by the culling front/back
surfaces and since they can be either clockwise or counter clockwise
there is no guaranteed vertex order that will work.

Add various settings to the CustomShader sections to allow the
rasterizer state to be overridden. Like the OM blend state, it will be
completely overridden if even a single option is present.

A simple option to disable front/back face culling would be:

[CustomShaderFoo]
cull = none

This could also be used to enable wireframe mode for debugging or change
any of the other state that one might need for whatever reason.

This also adds several helper functions to parse various basic types
from the Ini file, logging them if successful and logging a warning if
not - most of the file has not been converted yet, but many of the
options would probably benefit from switching.