CustomShader / CustomResource: Check for device changes
This should fix a theoretical issue where a game might create a device
during init and present at least one frame on it, then later switch to a
different device. If that were to happen we could potentially
substantiate custom resources and shaders, which would hold objects
associated with the old device that could cause problems if we attempt
to use them on the new device.

I was thinking this might be the cause of some glitches in Mad Max that
only happen if a CustomShader is set on the present call when the game
is launched (but works fine if added later and the config reloaded), but
no dice unfortunately...