Custom Shaders: Save and restore all RTVs, DSVs and UAVs
While it is possible to save the state of these with arbitrary resource
copying, doing so in a general way it actually a little tricky,
especially if UAVs are involved.

To make things easier, save them all off when a custom shader is run and
restore them on the way out.