Resource copying: Add a no_view_cache option to fix glitch in Mad Max
It seems that the mere existence of a view can cause glitches in some
situations. This occurs in Mad Max if a custom shader is called from the
present call that assigns the back buffer to a render target, which will
cause 3DMigoto to create a render target view of the back buffer and
hold onto a cached version.

If the cached render target view exists at some early point in Mad Max'
initialisation it glitches out the graphics for the whole game. Add a
no_view_cache option to release the view after it has been bound to the
pipeline (and therefore destroyed when it is later unbound, such as at
the end of the CustomShader section), which solves the glitches.