Resource Copying: Fix uninitialised flags when creating DSV
I noticed that backing up depth/stencil targets and restoring them was
failing to create a new view on the restore, which was due to the Flags
field being uninitialised.

For now set flags to be read/write as the most common option - later we
might fetch that from the old view (if applicable) or allow a keyword to
set it read only.