One last tweak on FC4 GS shader.
The input was line not triangle, this fixes that, and thus the declaration section in ASM is now identical.

This also seems to be quite close with regard to the array indexing.

Can't try it out now, but I think this is very close, maybe good enough to actuallly run correctly.