Replace the pathetic wsprintf with swprintf_s.
Microsoft has some lame software, but this is waaay up there.  wsprintf would crash with bad ptr.

When I changed it from %08lx%08lx to %016lx, it no longer worked.  Apparently wsprintf does not support that format, even though it's documented.  In any case, I'd expect it to print a string with no subsitution, not crash.

Anyway, replace every use of wsprintf as an archaic Microsoft variant we don't need.

This is probably why Chiri used the %08lx%08lx with hash>>32 to begin with.

I just don't need five versions of sprintf.  I need one, that works properly.

Also tweak a few local variables to clarify their scope.  We have a lot of inline declarations that are sub-optimal.  And found one bug because of a handle redefinition.