Pass shader path into D3DCompile so that #include can work reliably
If a shader uses #include, it currently requires that the path be
relative to the current working directory. In many cases that will be
where the d3d11.dll is located, but can differ. In Far Cry 4 & Far Cry
Primal, it changes depending on whether the UPlay or Steam versions of
the game is used.

The standard include handler will process included files relative to the
path of the shader being processed, which could be used to make these
consistent. In order to make that work, pass the path of the shader
into the D3DCompile routine.

Also change '((ID3DInclude*)(UINT_PTR)1)' to the equivalent
D3D_COMPILE_STANDARD_FILE_INCLUDE macro.