nvapi: Add unlock_convergence option
The nvapi wrapper in 3DMigoto will always unlock convergence, however
this is not always desirable. Some 3DVision enabled games may do
something sensible with convergence and we may not always want to block
that, but more importantly, this may interfere with Helix Mod if the DLL
is installed alongside that, as may happen for games that have both DX9
and DX11 renderers.

This comes about from the convergence remapping code, but does not seem
like it would have been intentional, however without fully understanding
all the nuances of that code I have opted to leave that behaviour in
place, but only when remapping is enabled in the d3dx.ini. If remapping
is not enabled, unlock_convergence=1 must now be specified.