Custom shaders: Add max_executions_per_frame
There are cases where we want to limit the number of times a custom
shader can be run in a frame even if it's trigger is invoked multiple
times. An example of this would be performing some processing prior to
any UI shaders to e.g. downsample a depth buffer or record certain
information in a buffer that will be used by further shaders.

This adds a max_executions_per_frame option to custom shaders to achieve
this, much like the max_copies_per_frame option in custom resources.