Add means to get mouse cursor visibility and position
This adds the ability to pass the cursor visibility and position into a
shader. Knowing that the cursor is visible can be used as an indication
that a HUD element should not be adjusted so that it lines up with the
hardware cursor.

This also gives us another piece we will need to implement a software
cursor, since the visibility can indicate when a software cursor needs
to be drawn and the position can be used to place the cursor.

A full blown software mouse still needs a few more pieces, namely the
ability to suppress showing the hardware mouse, as well as passing the
cursor in as a texture and ideally the cursor hotspot.

The mouse position is passed in in screen X & Y coordinates, which can
easily be scaled to whatever is required, e.g. to implement a 3D mouse
cursor using my crosshair.hlsl one could do something like:

[CustomShaderSoftwareMouse]
vs = ...
ps = ...
x1 = cursor_screen_x
y1 = cursor_screen_y
z1 = res_width
w1 = res_height
...
draw = 6, 0

 #include <crosshair.hlsl>
...
float4 params1 = IniParams.Load(int2(1, 0));
float2 cursor = (params1.xy / params1.zw) * 2 - 1;
o0.x += adjust_from_depth_buffer(cursor.x, cursor.y);

It's not clear that the "screen position" of the mouse is necessarily
the best units in all circumstances (e.g. if the game was not full
screen it would need to take the window position into account) and may
ultimately need to be improved, but for now it should cover most needs.