Manually fix bad shader constant. Example.
This is one of those stray bad constants.  Looks like the root cause is that the constant here is l(0,0,0x3f80000,0x3f80000) and the first two are not hex values, which breaks the sscan_f, putting it to the int case, which generates the 1605335.. numbers.

Since we know the constants are printed wrong anyway with %f default by the disassembler, the best fix would be to read the values from the James-Jones binary.