Helifax - Witcher 3 update to 1.21 patch
This is STILL based on the older version of 3DMigoto.
Main reasons:
- New version(s) don't compile properly the ASM shaders (so they don't
work)
- HUGE FPS drop (from 45 to 10 - 2x980TI in Surround) when using the
stereo2mono to get the "Depth-buffer"

Tested and this version works perfectly fine on the base game +
expansion on patch 1.21. It is possible I missed a few shaders?!?!

Re-fixed:
- Decals (quite a lot of them)
- Rain splashes /drops
- Blood
- Hairworks shadows
- Fixed witcher senses bubbles & warp bubbles so they now are at proper
depth.
- Only CROSSHAIR is at dynamic depth by default. Rest of the UI has code
but the FPS drop is big. So, disabled the code to pass the depth buffer
in the shaders (ini file).