Resource creation: Fix crash when creating a custom buffer
When a custom resource was created, 3DMigoto would crash because the
format used to create the resource was not propagated through to the
resource copying, and it had no other way to determine the format since
it is not stored in the buffer description, nor was there a view to find
it from. Propagate the format through if it is specified to fix the crash.