Add D3D11CreateDevice pFeatureLevel override.
There are finally a couple of games that error out when we return E_INVALIDARG for any non-DX11 requests.

This adds a d3dx.ini setting in System section, of allow_create_device, to allow the normal values CreateDevice and CreateDeviceAndSwapChain.

Two options, one to force it always to DX11, one to not change it at all.

Default is still to check DX11 and error on all others, as that has been very successful at fixing problems at launch.