Fix bugs for cb0[r19.w + 63].xyzw type references in arrays.
With large arrays like this in JC3, Globals[95], mapped to cb0, we would previously remove the 63 offset, because it was assumed that would map back to an actual named variable.

With arrays, that's not true, so we'd break array references.

This fix looks for the named variable in mCBufferData, like Globals[63] when used in applySwizzle, and if the index is non-zero, adds that reference back.  For a regular use case, it will be zero, like PointLights[0].

Not positive this won't break some other uses, but it definitely fixes this serious array indexing error.