Add skip_dxgi_device hooking option
MGSV crashes on Windows 10 with the anniversary update installed. This
seems to be due to the game rejecting the wrapped DXGIDevice like it did
the other wrapped devices, though it is a bit bizarre that it only
affects this specific OS configuration. Add a simple workaround to skip
wrapping that object to work around the crash. This does not seem to
have any negative effects on this game, but if necessary we could
alternatively solve the issue by adding a hooking version of the
DXGIDevice as well.