More hand tuning of JC3 shadow shader.
This continues that path of tweaking the code to use int/uint where clearly correct, and allow me to generate identical ASM to the original.

Not sure it fixes the crash, but this is very close to the original, outside of the optimizing compiler moving some blocks around, and register numbers changing.

This hand-tweak removes ALL of the ftoi and itof stuff that we normally have inserted in the code.