Only record stats if dump_usage is enabled
In the early stages of Firewatch (e.g. in the garage), hunting=1 causes
a rapid memory leak that exhausts all memory in about a minute before
Windows starts freaking out and the game crashes, while hunting=2 does
not cause this.

The stat collection seems to be the culprit as it is known to be able to
grow unbounded by it's nature (although it is surprising if it alone
could grow that fast), so make it conditional on dump_usage=1 since we
don't need stat collection if that is disabled.