Add enable_platform_update option to enable D3D11Device1
Dishonored 2 is the first known game that lacks a fallback path for
systems without the Windows 7 platform update, so we will need to
re-enable the old D3D11Device1 / D3D11Context1 code to use 3DMigoto with
it.

This commit adds an option to the d3dx.ini to enable getting a
D3D11Device1 from D3D11Device::QueryInterface and wrapping it and the
corresponding D3D11Context1. It is not tested, and I fully expect there
to be more issues lying in wait, but this will give the community a
starting point for more testing.