Fix incorrectly casting R11G11B10_FLOAT to R10G10B10A2_TYPELESS
DHR reported in Hitman attempting to use stereo2mono to swap the eyes
for the reflection failed with an error like:

Resource copy CreateCompatibleView failed: 80070057
  View Type = Shader Resource
    Format = R11G11B10_FLOAT (26)
    ViewDimension = TEXTURE2D
      Texture2D.MostDetailedMip = 0
      Texture2D.MipLevels = -1
  Resource Type = Texture2D
    Width = 3840
    Height = 1080
    MipLevels = 1
    ArraySize = 1
    Format = R10G10B10A2_TYPELESS (23)
    SampleDesc.Count = 1
    SampleDesc.Quality = 0
    Usage = 0
    BindFlags = 0x8
    CPUAccessFlags = 0x0
    MiscFlags = 0x0

The problem is that the stereo2mono created in intermediate typeless
resource, but the format was not compatible with the original resource,
nor the format it later attempted to use to create the new view. There
is no appropriate typeless variant of this format, so fix the problem by
never attempting to convert it to a typeless format.

Reported-by: DHR