Add allow_dxgi1_2=2 to allow unwrapped DXGIFactory2 for the Steam Overlay
The Steam overlay does not work in some games including The Witcher 3 on
Windows 7/8 (but works fine in other games such as Firewatch). This
seems to be related to rejecting queries for a DXGIFactory2 object, but
also fails when we return a wrapped factory 2 object.

The log does not show any calls on the wrapped factory 2 object,
suggesting that Steam is somehow rejecting it outright. It is not clear
exactly why this may be, but for now add allow_dxgi1_2=2 to allow the
original unwrapped DXGIFactory2 object to be passed back to the overlay.
Naturally this may be an issue for games that legitimately use this
object since we might not wrap the present call, but we don't care when
it is just being used for the Steam overlay.