Log game's driver profile - first pass
This logs the current profile for the game, as well as the base and
global profiles to the d3d11_log.txt using a format that is very similar
to Geforce Profile Manager. The eventual goal will be to give 3DMigoto a
means to check if the profile is correct and eventually to set them
directly (depending on what we do about the admin privilege
requirement).

The default stereo profile is not yet logged - the relevant function
isn't listed in the nvapi headers we are currently using, so we will
need to update that first.

It does not yet decode the internal settings.