Decode "internal" settings
This one is pretty involved - at a high level it is following the same
procedure to decode the internal settings that we have discovered
previously, and that is now implemented in nvidia profile inspector as
well as my cleanup script.

However, there was a significant challenge as the API does not seem to
expose any way to determine which settings are internal, and as that
varies for each profile it cannot be hard coded.

Inspired by the solution that nvidia profile inspector used, this calls
the undocumented NvAPI_DRS_SaveSettingsToFileEx to temporarily write the
settings out to a file formatted exactly as Geforce Profile Manager
would do, so that it can then read that file back in looking for which
settings have the internal flag. Of course, this temporary file suffers
exactly the same encoding woes as Geforce Profile Manager, so we have to
essentially treat it as a raw binary file while we pull out what we
need.

The end result is that the log file now gives us a more complete picture
of what is affecting the game, and brings us a step closer to allowing
3DMigoto to write profiles back.

Currently, this writes the temporary file for each profile that it logs,
so using it with log_all_profiles is a bad idea.