Create map of internal settings for all profiles
If we are logging multiple profiles we save the full set of profiles to
disk and read it back for each one. When we are only dumping the base
profile and the game profile that is not a big deal, but if we are
dumping every profile it can take minutes.

Dump the file once and parse the whole thing to create a map of every
profile to their internal settings, that we can reference as we dump out
each one. Clear the map once we are finished logging to recover memory
usage.