WIP to elevate our privileges in order to install a driver profile
The DRS API only allows us to save changes to a profile if we are
running with admin privileges. Since we are a dll loaded into another
application we have no control over whether we are run as admin or not,
and it is not a good idea to require that the user run the game as admin
since we don't know what consequences that might have.

This is work in progress of one possible way we might elevate our
privileges in order to install the driver profile. This spawns a rundll
process as admin that will load us and call a given entry point,
allowing us to run the minimum code necessary to install a profile.

The major problem found with this approach is that rundll is from a
bygone era where spaces didn't exist, and turns out it was never updated
to cope with them - rundll will silently fail if the path to the game
contains a space. Even using the shortname as recommended in the
documentation does not always help, as that may be disabled on some
drives, especially non-system drives, and even if it is enabled there is
a possibility that some directory may not have a shortname.

There may be other solutions we could implement - we could copy
ourselves to an alternate location that is less likely to have spaces
and more likely to have a shortname (e.g. the temp directory), or we
could write a small helper application that we distribute alongside
3DMigoto (or perhaps even embedded as a resource inside 3DMigoto that we
write to disk only when required).

Whether we do end up using the rundll approach or not, this still seems
like potentially useful code, so into the git history it goes.